fix(collisions): move narrow phase plugin to correct folder and use new system syntax
